config/requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
requests==2.32.3
pandas>=2.0.0,<2.3.0
pydantic>=2.0.0
beautifulsoup4==4.12.3
tiktoken==0.7.0
langchain==0.2.12
langchain-community==0.2.11
faiss-cpu>=1.9.0
openai==1.40.6
tenacity==8.3.0
boto3==1.34.0
botocore==1.34.0
PyPDF2==3.0.1
python-docx==1.1.0
reportlab==4.0.4
python-pptx==0.6.21
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
numpy>=1.24.0
